backend/requirements.txt
wheel
ipython==7.2.0
Django==2.0
psycopg2-binary==2.8.1
requests==2.21.0
factory-boy==2.11.1

django-extensions==2.1.4
djangorestframework==3.9.2
django-map-widgets==0.2.2
django-bootstrap4==0.0.7
django-extra-fields==1.1.0
ipdb
Fabric3==1.14.post1

celery==4.3.0
django-celery-beat==1.4.0
django-celery-results==1.0.4